Basic publish/unpublish behavior

+++ app/routes/posts/$postId.tsx
@@ -3,7 +3,7 @@ import { json, redirect } from "@remix-run/node";
 import { Form, useCatch, useLoaderData } from "@remix-run/react";
 import invariant from "tiny-invariant";
 
-import type { Post } from "~/models/post.server";
+import { Post, updatePost } from "~/models/post.server";
 import { deletePost } from "~/models/post.server";
 import { getPost } from "~/models/post.server";
 import { requireUserId } from "~/session.server";
@@ -28,35 +28,69 @@ export const action: ActionFunction = async ({ request, params }) => {
   const userId = await requireUserId(request);
   invariant(params.postId, "postId not found");
 
-  await deletePost({ userId, id: params.postId });
+  const formData = await request.formData();
+  const action = formData.get("action");
+  if (action === "publish") {
+    await updatePost({ userId, id: params.postId, published: true });
+    return redirect(`/posts/${params.postId}`);
+  } else if (action === "unpublish") {
+    await updatePost({ userId, id: params.postId, published: false });
+    return redirect(`/posts/${params.postId}`);
+  } else if (action === "delete") {
+    await deletePost({ userId, id: params.postId });
+    return redirect("/");
+  }
+};
 
-  return redirect("/");
+const PostActions = ({ post }: { post: Post }) => {
+  return (
+    <Form method="post">
+      {!post.published ? (
+        <button
+          type="submit"
+          name="action"
+          value="publish"
+          className="mx-4 rounded bg-blue-500 py-2 px-4 text-white hover:bg-blue-600 focus:bg-blue-400"
+        >
+          Publish
+        </button>
+      ) : (
+        <button
+          type="submit"
+          name="action"
+          value="unpublish"
+          className="mx-4 rounded bg-blue-500 py-2 px-4 text-white hover:bg-blue-600 focus:bg-blue-400"
+        >
+          Unpublish
+        </button>
+      )}
+      <button
+        type="submit"
+        name="action"
+        value="delete"
+        className="rounded bg-red-500  py-2 px-4 text-white hover:bg-blue-600 focus:bg-blue-400"
+      >
+        Delete
+      </button>
+    </Form>
+  );
 };
 
 export default function PostDetailsPage() {
-  const data = useLoaderData() as LoaderData;
+  const { post } = useLoaderData() as LoaderData;
   const user = useUser();
 
   return (
     <div className="post">
-      <h2>{data.post.title}</h2>
-      {!data.post.published && (
+      <h2>{post.title}</h2>
+      {!post.published && (
         <div className="py-6">
           <small>This post has not yet been published.</small>
         </div>
       )}
-      <p>{data.post.content}</p>
+      <p>{post.content}</p>
       <hr className="my-4" />
-      {data.post.authorId === user.id && (
-        <Form method="post">
-          <button
-            type="submit"
-            className="rounded bg-blue-500  py-2 px-4 text-white hover:bg-blue-600 focus:bg-blue-400"
-          >
-            Delete
-          </button>
-        </Form>
-      )}
+      {post.authorId === user.id && <PostActions post={post} />}
     </div>
   );
 }
